backend/requirements.txt
feedparser>=6.0
requests>=2.31
google-genai>=1.0
pyyaml>=6.0
jinja2>=3.1
python-dotenv>=1.0
beautifulsoup4>=4.12
numpy>=1.24
supabase>=2.0
python-slugify>=8.0
resend>=1.0
